Move code to src and auxiliary methods to an utils file
